Implements (most of) the remainder of the module.
There's one remaining function to implement that `applyPatches`. However,in
addition to that, it turns out that most functions can take a callback in their
options to run things async'ly. Unfortunately, it's not entirely documented
which functions can take a callback. This can be done at a later date. As of
now, the library is quite usable.